Safari-609.1.11.1/LayoutTests/imported/w3c/web-platform-tests/service-workers/service-worker/fetch-event-respond-with-response-body-with-invalid-chunk.https-expected.txt

CONSOLE MESSAGE: FetchEvent.respondWith received an error: ReadableStream chunk enqueueing in the sink failed

PASS Response with a ReadableStream having non-Uint8Array chunks should be transferred as errored 
